# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "afrozsatvilkar2014@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "summary": "register",
                "description": "register the user for login",
                "operationId": "2f25885f3b6f61ca11724ec302d2e439",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "login",
                "description": "user login",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "login Success"
                    },
                    "401": {
                        "description": "we can not find the user with that e-mail address You need to register first"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "logout",
                "description": "logout user",
                "operationId": "67d5147477c548bfbc3f982833bbc7bd",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "summary": "profile",
                "description": "user profile",
                "operationId": "4373251f8140344279ac32a1ac817cef",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/verifyemail/{token}": {
            "get": {
                "summary": "verify email",
                "description": "user email verify",
                "operationId": "1a38ac7b4cc5080891462bc89fce0b61",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Email is Successfully verified"
                    },
                    "202": {
                        "description": "Email Already verified"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}